(HP-1796) Moved AddButton to own folder
